moved circle() and boundary() to gui.utils
