+++ suite2p/gui/merge.py
@@ -2,9 +2,9 @@ import numpy as np
 import pyqtgraph as pg
 from scipy import stats
 from PyQt5 import QtGui
-from .. import utils
 from .. import extraction
 from . import masks
+from . import utils
 
 def distance_matrix(parent, ilist):
     idist = 1e6 * np.ones((len(ilist), len(ilist)))
