+++ suite2p/gui/utils.py
@@ -1,0 +1,23 @@
+import numpy as np
+
+
+def boundary(ypix,xpix):
+    """ returns pixels of mask that are on the exterior of the mask """
+    ypix = np.expand_dims(ypix.flatten(),axis=1)
+    xpix = np.expand_dims(xpix.flatten(),axis=1)
+    npix = ypix.shape[0]
+    idist = ((ypix - ypix.transpose())**2 + (xpix - xpix.transpose())**2)
+    idist[np.arange(0,npix),np.arange(0,npix)] = 500
+    nneigh = (idist==1).sum(axis=1) # number of neighbors of each point
+    iext = (nneigh<4).flatten()
+    return iext
+
+
+def circle(med, r):
+    """ returns pixels of circle with radius 1.25x radius of cell (r) """
+    theta = np.linspace(0.0,2*np.pi,100)
+    x = r*1.25 * np.cos(theta) + med[0]
+    y = r*1.25 * np.sin(theta) + med[1]
+    x = x.astype(np.int32)
+    y = y.astype(np.int32)
+    return x,y
